refactor: rename EbillEndpoint to EbillsEndpoint and EmailEndpoint to EmailsEndpoint
